models/requirements.txt
aiobotocore==1.1.2
aiohttp==3.7.3
aioitertools==0.7.1
alembic==1.4.2
anyio==3.6.1
apispec==1.3.3
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.11.1
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.5
awscli==1.18.142
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
boto3==1.16.43
botocore==1.17.44
bs4==0.0.1
cached-property==1.5.1
cachetools==4.1.1
cattrs==0.9.2
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
contextlib2==21.6.0
contextvars==2.4
croniter==0.3.34
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distlib==0.3.1
Django==3.0.4
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
entrypoints==0.3
fastapi==0.78.0
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
fsspec==0.8.5
funcsigs==1.0.2
future==0.18.2
google==3.0.0
google-auth==1.4.2
google-colab==1.0.0
graphviz==0.14
gunicorn==19.10.0
h11==0.13.0
idna==2.8
idna-ssl==1.1.0
image-slicer==2.1.1
immutables==0.18
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.1.1
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.3
jmespath==0.10.0
json-merge-patch==0.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==21.4.0
kiwisolver==1.2.0
lab==6.1
lazy-object-proxy==1.5.0
lockfile==0.12.2
lxml==4.6.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.2
mistune==0.8.4
multidict==5.1.0
natsort==7.0.1
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==5.2.2
numpy==1.19.0
oauthlib==3.1.0
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.1
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
pkginfo==1.6.1
pluggy==0.13.1
portpicker==1.2.0
prison==0.1.3
prometheus-client==0.8.0
prompt-toolkit==1.0.18
psutil==5.7.0
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.9.1
Pygments==2.6.1
PyJWT==1.7.1
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.2
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python3-openid==3.2.0
pytz==2019.3
pytzdata==2019.3
pyxlsb==1.0.6
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
readme-renderer==28.0
redis==3.5.3
redislite==6.0.674960
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.5
s3fs==0.5.1
s3transfer==0.3.3
Send2Trash==1.5.0
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.17.2
six==1.12.0
sniffio==1.2.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.1
starlette==0.19.1
structlog==21.5.0
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.2
thrift==0.13.0
toml==0.10.2
tornado==4.5.3
tqdm==4.51.0
traitlets==4.3.3
twine==3.2.0
txt2tags==3.7
typing==3.7.4.1
typing-extensions==4.1.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
uuid==1.30
uvicorn==0.16.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
xlrd==1.2.0
yarl==1.6.3
zipp==3.1.0
zope.deprecation==4.4.0
